requirements.txt
allure-pytest==2.8.6
allure-python-commons==2.8.6
appdirs==1.4.3
atomicwrites==1.4.0
attrs==19.3.0
beautifulsoup4==4.9.0
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cssselect==1.1.0
fake-useragent==0.1.11
Flask==1.0.2
idna==2.8
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.10
lxml==4.5.0
MarkupSafe==1.1.1
more-itertools==8.2.0
parse==1.15.0
pluggy==0.13.1
py==1.8.1
pyee==7.0.1
PyMySQL==0.9.3
pyppeteer==0.0.25
pyquery==1.4.1
pytest==4.5.0
pytest-html==1.19.0
pytest-metadata==1.9.0
PyTestReport==0.2.1
PyYAML==5.1.2
requests==2.22.0
requests-html==0.10.0
requests-toolbelt==0.9.1
six==1.14.0
soupsieve==2.0
tqdm==4.46.0
urllib3==1.25.9
w3lib==1.21.0
wcwidth==0.1.9
websockets==8.0.2
Werkzeug==1.0.1
zipp==3.1.0
